[refactor]: read-together, room-together를 room으로 수정 (#37)
