--- src/main/java/com/example/bookjourneybackend/domain/readAlone/ReadAlone.java
@@ -1,4 +1,0 @@
-package com.example.bookjourneybackend.domain.readAlone;
-
-public class ReadAlone {
-}
