--- src/main/java/com/example/bookjourneybackend/domain/readTogether/domain/repository/ReadTogetherRepository.java
@@ -1,4 +1,0 @@
-package com.example.bookjourneybackend.domain.readTogether.domain.repository;
-
-public interface ReadTogetherRepository {
-}
