--- src/main/java/com/example/bookjourneybackend/domain/readTogether/dto/response/GetRoomInfoResponse.java
@@ -1,10 +1,0 @@
-package com.example.bookjourneybackend.domain.readTogether.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class GetRoomInfoResponse {
-    private RoomData roomData;
-}
