+++ src/main/java/com/example/bookjourneybackend/domain/room/controller/RoomController.java
@@ -1,4 +1,26 @@
 package com.example.bookjourneybackend.domain.room.controller;
 
+import com.example.bookjourneybackend.domain.room.dto.response.GetRoomInfoResponse;
+import com.example.bookjourneybackend.domain.room.service.RoomService;
+import com.example.bookjourneybackend.global.response.BaseResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/rooms")
 public class RoomController {
+
+    private final RoomService roomService;
+
+    @GetMapping("/{roomId}/info")
+    public BaseResponse<GetRoomInfoResponse> getRoomInfo(@PathVariable Long roomId) {
+
+        return BaseResponse.ok(roomService.getRoomInfo(roomId));
+    }
 }
