+++ src/main/java/com/example/bookjourneybackend/domain/room/domain/Room.java
@@ -1,63 +1,90 @@
 package com.example.bookjourneybackend.domain.room.domain;
 
-import com.example.bookjourneybackend.domain.book.domain.Book;
-import com.example.bookjourneybackend.domain.readTogether.domain.ReadTogether;
+import com.example.bookjourneybackend.domain.favorite.domain.Favorite;
+import com.example.bookjourneybackend.domain.record.domain.Record;
 import com.example.bookjourneybackend.domain.user.domain.User;
+import com.example.bookjourneybackend.domain.userRoom.domain.UserRoom;
 import com.example.bookjourneybackend.global.entity.BaseEntity;
 import jakarta.persistence.*;
-
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity
-@Table(name = "rooms")
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
 @NoArgsConstructor
+@Entity
 @Getter
+@Table(name = "rooms")
 public class Room extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long roomId;
 
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private RoomType roomType;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private UserRole userRole;
-
-    @Column(nullable = false)
-    private Double userPercentage;
-
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
+    @JoinColumn(name = "user_room_id", nullable = false)
+    private UserRoom userRoom;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "book_id", nullable = false)
-    private Book book;
+    @Column(nullable = false, length = 60)
+    private String roomName;
 
     @Column(nullable = false)
-    private Integer currentPage;
+    private boolean isPublic;
 
-    // ReadTogether와의 관계 추가
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "read_together_id", nullable = false)
-    private ReadTogether readTogether;
+    private Integer password;
 
+    @Column(nullable = false)
+    private Double roomPercentage;
 
+    @Column(nullable = false)
+    private LocalDateTime progressStartDate;
+
+    @Column(nullable = false)
+    private LocalDateTime progressEndDate;
+
+    @Column(nullable = false)
+    private LocalDateTime recruitStartDate;
+
+    @Column(nullable = false)
+    private LocalDateTime recruitEndDate;
+
+    @Column(nullable = false)
+    private Integer recruitCount;
+
+    @Column(nullable = false)
+    private Integer recordCount;
+
+    @OneToMany(mappedBy = "room", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<UserRoom> userRooms = new ArrayList<>();
+
+    @OneToMany(mappedBy = "room", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Record> records = new ArrayList<>();
 
     @Builder
-    public Room(Long roomId, RoomType roomType, UserRole userRole, Double userPercentage, User user, Book book,Integer currentPage, ReadTogether readTogether) {
+    public Room(Long roomId, UserRoom userRoom, String roomName, boolean isPublic, Integer password, Double roomPercentage, LocalDateTime progressStartDate, LocalDateTime progressEndDate, LocalDateTime recruitStartDate, LocalDateTime recruitEndDate, Integer recruitCount, Integer recordCount) {
         this.roomId = roomId;
-        this.roomType = roomType;
-        this.userRole = userRole;
-        this.userPercentage = userPercentage;
-        this.user = user;
-        this.book = book;
-        this.currentPage = currentPage;
-        this.readTogether = readTogether;
+        this.userRoom = userRoom;
+        this.roomName = roomName;
+        this.isPublic = isPublic;
+        this.password = password;
+        this.roomPercentage = roomPercentage;
+        this.progressStartDate = progressStartDate;
+        this.progressEndDate = progressEndDate;
+        this.recruitStartDate = recruitStartDate;
+        this.recruitEndDate = recruitEndDate;
+        this.recruitCount = recruitCount;
+        this.recordCount = recordCount;
+    }
+
+    public void addUserRoom(UserRoom userRoom) {
+        this.userRooms.add(userRoom);
+        userRoom.setRoom(this);
+    }
+
+    public void addRecord(Record record) {
+        this.records.add(record);
     }
 }
