+++ src/main/java/com/example/bookjourneybackend/domain/userRoom/service/RoomService.java
@@ -1,0 +1,4 @@
+package com.example.bookjourneybackend.domain.userRoom.service;
+
+public class RoomService {
+}
